add remove tasks

+++ 18-Projects/alitaskmanager/tpl/tpl-index.php
@@ -26,14 +26,16 @@
         <div class="menu">
           <div class="title">Folders</div>
           <ul class="folder-list">
+          <li class="<?= isset($_GET['folder_id'])? '' : 'active' ?>"><i class="fa fa-folder-open"></i>All</li>
+         
             <?php foreach ($folders as $folder) : ?>
-              <li>
+              <li class="<?= ($_GET['folder_id'] == $folder->id) ? 'active' : '' ?>">
                 <a href="?folder_id=<?= $folder->id ?>"><i class="fa fa-folder"></i><?= $folder->name ?></a>
-                <a href="?delete_folder=<?= $folder->id ?>"><i class="fa fa-trash-o"></i></a>
+                <a href="?delete_folder=<?= $folder->id ?>"><i class="fa fa-trash-o" onclick="return confirm('Are You Sure To Delete Folder <?= $folder->name ?>?')"></i></a>
 
               </li>
             <?php endforeach; ?>
-            <li class="active"><i class="fa fa-folder-open"></i>Current Folder</li>
+           
           </ul>
         </div>
         <div>
@@ -43,40 +45,47 @@
       </div>
       <div class="view">
         <div class="viewHeader">
-          <div class="title">Manage Tasks</div>
+          <div class="title">
+            <input type="text" id="taskNameInput" style="width: 65%; line-height:30px; margin-left:3%"  placeholder="Add New Task">
+           
+
+          </div>
           <div class="functions">
             <div class="button active">Add New Task</div>
             <div class="button">Completed</div>
             <div class="button inverz"><i class="fa fa-trash-o"></i></div>
           </div>
         </div>
-        <div class="content">
+        <di
+        v class="content">
           <div class="list">
             <div class="title">Today</div>
             <ul>
-              <li class="checked"><i class="fa fa-check-square-o"></i><span>Update team page</span>
+              <?php if(sizeof($tasks) > 0): ?>
+              <?php foreach($tasks as $task):  ?>
+                
+              <li class="<?= $task->is_done? "checked" : ""; ?>">
+              <!-- ternary oprator for is_done task or not -->
+              <i class="<?php echo $task->is_done? "fa fa-check-square-o" : "fa fa-square-o" ?>"></i>
+              
+              <span><?= $task->title ?></span>
+
                 <div class="info">
-                  <div class="button green">In progress</div><span>Complete by 25/04/2014</span>
+                  <span class="created-at">Created At<?= $task->created_at ?></span>
+                  <a href="?delete_task=<?= $task->id ?>"  class="fa fa-trash-o" onclick="return confirm('Are You Sure To Delete Task <?= $task->title ?> ?')"></a>
                 </div>
               </li>
-              <li><i class="fa fa-square-o"></i><span>Design a new logo</span>
-                <div class="info">
-                  <div class="button">Pending</div><span>Complete by 10/04/2014</span>
-                </div>
-              </li>
-              <li><i class="fa fa-square-o"></i><span>Find a front end developer</span>
-                <div class="info"></div>
-              </li>
+
+              <?php endforeach; ?>
+              <?php else: ?>
+                <li>No Task Here...</li>
+              <?php endif; ?>
+                        
+              
+
             </ul>
           </div>
-          <div class="list">
-            <div class="title">Tomorrow</div>
-            <ul>
-              <li><i class="fa fa-square-o"></i><span>Find front end developer</span>
-                <div class="info"></div>
-              </li>
-            </ul>
-          </div>
+         
         </div>
       </div>
     </div>
